exit 1

+++ images/linux/scripts/base/apt-mock.sh
@@ -5,7 +5,11 @@ set -x
 prefix=/usr/local/bin
 
 for tool in apt apt-get apt-fast apt-key;do
-  which $tool || continue
+  which $tool
+  if [ $? -e 0 ];then
+    echo "exit $tool"
+    exit 1
+  fi
   real_tool=`which $tool`
   cat >$prefix/$tool <<EOT
 #!/bin/sh
